Tensorrt (#151)
* feat: add nvidia NX support

* feat: improve libtorch install

* feat(all): adaption of apollo on cuda 12+trt10 (#57)

* feat(all): adaption of apollo on cuda 12+trt10

* fix: fix build

* fix: remove vtk

* feat: remove paddle deps

* fix: remove vtk

* fix: fix ncut error

* feat: update docker due to miss lib/fastrtps

* chore: improve image processor

* fix: check tensorrt

---------


* fix: fix tensorrt build by adding rules_cuda dependency

* fix: fix build error of tensorrt

* fix: fix build

* fix: fix decode_video api to 7.0

* fix: fix tensorrt build error

* fix: fix tensorrt reshape dims

* fix(perception): mutiple issues of tensorrt 10 adaptation

* sync from main (#181)

* feat: add gnss conf tool (#153)

* feat: add wheelos bazel registry url (#161)

* 166 feature independent dreamview frontend (#168)

* feat: Independent dreamview frontend

* chore: remove frontend directory from git tracking

* fix: fix dkit protocol (#134)

* fix: fix dkit protocol

* fix: call can_sender_->Update() in Emergency()

* feat: add remote debugging tool (#178)

* 177 feature remote debugging tool (#179)

* feat: add remote debugging tool

* chore: improve whl-remote

* feat: Add auto-start feature (#84)

* feat: Add auto-start feature

* chore: Optimized implementation

* fix: fix start/enter issues of whl.sh


* chore(docker/scripts/whl.sh): support custom names and tags


---------


---------


* feat(perception) benchmark tool for lidar detection

* fix(perception): issue of lidar benchmark

* fix(perception): spatio_temporal_ground_detector proto missing fields

* fix: fix compilation issue


* chore: replace local_config_cuda with rules_cuda

- replace local_config_cuda with rules_cuda
- add cudnn and tenesorrt extension


* chore: change opencv from third_party to bazel_dep


* fix(perception): explicit batch on trt10 branch

* fix(perception): plugin of tensorrt support nchw input

* fix(perception): add batchsize check for plugins of rcnn and rpn

* fix(cyber): fix segmentation fault issue of exiting


* fix(drivers/camera): fix compilation issue with opencv


* add centerpoint trt inference

* fix(cyber): core when reading tls variable from inline function on jetpack 6.2 with specify O2 optimization

* chore(perception): update config of centerpoint with trt inference

* chore(perception): add centertrt define on pipeline

* chore(docker): add user to video group on aarch64

* chore(perception): unused variable fixed

* fix(whl.sh): logic of checking gpu

---------

+++ modules/perception/tool/benchmark/lidar/loader/async_sequence_data_loader.h
@@ -21,6 +21,7 @@
 #include <mutex>
 #include <string>
 #include <vector>
+
 #include "modules/perception/tool/benchmark/lidar/ctpl/ctpl.h"
 #include "modules/perception/tool/benchmark/lidar/loader/sequence_data_loader.h"
 
@@ -125,7 +126,7 @@ bool AsyncSequenceDataLoader<DataType>::query_next(
     load_success = iter->second->load_success;
   }
   // prefetch next data
-  for (int i = _idx + 1;
+  for (size_t i = _idx + 1;
        i <= std::min(static_cast<std::size_t>(_idx) + _prefetch_data_size,
                      _filenames[0].size() - 1);
        ++i) {
